Upgrade to Gradle 6.5.1
Ran ./gradlew wrapper --gradle-version 6.5.1 to upgrade the
reference to gradle. Upgraded the Dockerfile to use the correct
version of the Gradle as well.  I was able to build the jar
locally as well as a new Docker image without any issues.

In addition I added a check to verify that you running with the
proper version of Java. When I was first trying to get things
to compile gradle would fail with a cryptic message about plugins
because I was using a earlier version of Java. Hopefully this
will prevent that from happening in the future.